Extract import from editor to Xyz.ts
