+++ sashafirsov/src/xyz/Editor.mock.ts
@@ -1,0 +1,22 @@
+export const TwoFramesMock =                        `6
+Created by chemcoord http://chemcoord.readthedocs.io/en/latest/
+O 0.000000 0.000000  0.000000
+H 0.758602 0.000000  0.504284
+H 0.260455 0.000000 -0.872893
+O 3.000000 0.500000  0.000000
+H 3.758602 0.500000  0.504284
+H 3.260455 0.500000 -0.872893
+11
+# pyridine molecule 
+C       -0.180226841      0.360945118     -1.120304970
+C       -0.180226841      1.559292118     -0.407860970
+C       -0.180226841      1.503191118      0.986935030
+N       -0.180226841      0.360945118      1.29018350
+C       -0.180226841     -0.781300882      0.986935030
+C       -0.180226841     -0.837401882     -0.407860970
+H       -0.180226841      0.360945118     -2.206546970
+H       -0.180226841      2.517950118     -0.917077970
+H       -0.180226841      2.421289118      1.572099030
+H       -0.180226841     -1.699398882      1.572099030
+H       -0.180226841     -1.796059882     -0.917077970
+`;
