+++ sashafirsov/src/xyz/Editor.tsx
@@ -6,15 +6,15 @@ import styles from './Editor.module.css';
 
 // from https://github.com/microsoft/monaco-editor/blob/main/samples/browser-esm-vite-react/src/main.tsx
 
-import Xyz, { ElementXyz, XyzSlide } from '../xyz/Xyz';
-import * as console from 'console';
+import Xyz from '../xyz/Xyz';
+import { TwoFramesMock } from './Editor.mock';
 
 export const Editor: VFC = () => {
-    const [drawing, saveDrawing] = useLocalStorage<Xyz>('xyzdrawing');
+    const [, shareDrawing] = useLocalStorage<Xyz>('xyzdrawing');
 
     const [editor, setEditor] = useState<monaco.editor.IStandaloneCodeEditor | null>(null);
     const monacoEl = useRef(null);
-    const [_selection, setSelection] = useLocalStorage('EditorSelection', [0, 0]);
+    const [, shareSelection] = useLocalStorage('EditorSelection', [0, 0]);
 
     useEffect(() => {
         if (monacoEl) {
@@ -24,63 +24,20 @@ export const Editor: VFC = () => {
                 const mEditor = monaco.editor.create(monacoEl.current!, {
                     automaticLayout: true,
                     readOnly: false,
-                    value:
-                        `6
-Created by chemcoord http://chemcoord.readthedocs.io/en/latest/
-O 0.000000 0.000000  0.000000
-H 0.758602 0.000000  0.504284
-H 0.260455 0.000000 -0.872893
-O 3.000000 0.500000  0.000000
-H 3.758602 0.500000  0.504284
-H 3.260455 0.500000 -0.872893
-11
-# pyridine molecule 
-C       -0.180226841      0.360945118     -1.120304970
-C       -0.180226841      1.559292118     -0.407860970
-C       -0.180226841      1.503191118      0.986935030
-N       -0.180226841      0.360945118      1.29018350
-C       -0.180226841     -0.781300882      0.986935030
-C       -0.180226841     -0.837401882     -0.407860970
-H       -0.180226841      0.360945118     -2.206546970
-H       -0.180226841      2.517950118     -0.917077970
-H       -0.180226841      2.421289118      1.572099030
-H       -0.180226841     -1.699398882      1.572099030
-H       -0.180226841     -1.796059882     -0.917077970
-`,
+                    value: TwoFramesMock,
                     language: 'python'
                 });
 
                 function onEditChanged() {
-                    const xyz = new Xyz();
-
-                    const lines = mEditor.getValue().split('\n');//.filter(t => t.trim());
-
-                    for (let i = 0; i < lines.length; i++) {
-                        const line = lines[i].trim();
-                        if (!line || line.startsWith('#')) // ignore blank lines and comments
-                            continue;
-
-                        const n = parseInt(lines[i++]);
-                        const comment = lines[i++];
-                        const elements = lines.slice(i, i + n).map((line, k) => {
-                            const [element, x, y, z] = line.split(/\s+/);
-                            return new ElementXyz(element, parseFloat(x), parseFloat(y), parseFloat(z), i + k + 1);
-                        });
-                        const slide = new XyzSlide();
-                        slide.comment = comment;
-                        slide.elements = elements;
-                        xyz.addSlide(slide);
-                        i += n - 1;
-                    }
-                    saveDrawing(xyz);
+                    setTimeout(() => shareDrawing(Xyz.parse(mEditor.getValue())), 10);
                 }
 
                 mEditor.onDidChangeModelContent(onEditChanged);
                 onEditChanged();
                 mEditor.onDidChangeCursorSelection(e => {
-                    setSelection([e.selection.startLineNumber, e.selection.endLineNumber]);
+                    shareSelection([e.selection.startLineNumber, e.selection.endLineNumber]);
                 });
-                // @ts-expect-error always available in browser
+                // @ts-expect-error document always available in browser
                 document.getElementById('import-file').addEventListener(
                     'change',
                     () => {
@@ -90,31 +47,9 @@ H       -0.180226841     -1.796059882     -0.917077970
                             for (const file of fileInput.files) {
                                 const reader = new FileReader();
                                 reader.onload = (e) => {
-                                    if (file.name.endsWith('.poscar')) {
-                                        // convert poscar to xyz
-
-                                        const contents = e.target?.result as string;
-                                        const [comment, scaling, a1, a2, a3, ionSpecies, ionNumbers, directOrCartesian, ...positions] = contents.split('\n');
-
-                                        const species = ionSpecies.trim().split(/\s+/);
-                                        const speciesCount = ionNumbers.trim().split(/\s+/).map(n => parseInt(n));
-                                        const xyzLines = [
-                                            '###########',
-                                            '# ' + file.name,
-                                            '# ' + comment,
-                                            '# scaling ' + scaling,
-                                            speciesCount.reduce((total, v) => total + v, 0)];
-                                        let pi = 0;
-                                        for (let si = 0; si < species.length; si++) {
-                                            for (let sci = 0; sci < speciesCount[si]; sci++, pi++)
-                                                xyzLines.push(`${species[si]} ${positions[pi]}`);
-                                        }
-                                        mEditor.setValue(mEditor.getValue() + xyzLines.join('\n'));
-                                    } else {
-                                        // no need to convert
-                                        const contents = e.target?.result as string;
-                                        mEditor.setValue(mEditor.getValue() + `\n###################\n# ${file.name}\n\n` + contents);
-                                    }
+                                    const contents = e.target?.result as string;
+                                    const text = file.name.endsWith('.poscar') ? Xyz.poscar2xyzString(contents) : contents;
+                                    mEditor.setValue(mEditor.getValue() + `\n###################\n# ${file.name}\n` + text);
                                 };
                                 reader.readAsText(file);
                             }
